==> notebooks/1. ETLS.ipynb <==
import pandas as pd
import numpy as np

# 1. ETLS

En este notbook se proporcionarán los pasos realizados para la limpieza y analisis de los datasets con los que vamos a partir para el desarrollo del proyecto. El fin es conseguir un dataset limpio y sencillo con el que poder jugar y así poder desarrollar nuestra idea de proyecto.

### 1.1 Centros médicos Madrid

centros_medicos = pd.read_csv("../data/centros_medicos_madrid.csv", encoding='latin-1', sep=';', error_bad_lines=False)
centros_medicos.head()

centros_medicos.shape

centros_medicos.keys()

centros_medicos.isnull().sum()

* Como vemos hay muchas columnas que tienen mucshísimos valores nulos y las vamos a tener que eliminar
* La eliminación de estas columnas no influye en absoluto con el desarrollo de nuestro proyecto porque esa información no es relevante para conseguir el fin de este trabajo
* Por lo tanto eliminaremos las siguientes diez columnas: 
    - DESCRIPCION-ENTIDAD
    - HORARIO
    - EQUIPAMIENTO
    - DESCRIPCION
    - PLANTA
    - PUERTA
    - ESCALERAS
    - ORIENTACION
    - FAX
    - EMAIL

centros_medicos = centros_medicos.drop(['DESCRIPCION-ENTIDAD', 'HORARIO', 'EQUIPAMIENTO', 'DESCRIPCION', 'PLANTA', 'PUERTA', 'ESCALERAS', 'ORIENTACION', 'FAX', 'EMAIL'], axis=1)
centros_medicos.head()

centros_medicos.isnull().sum()

centros_medicos.info()

#### Estudio profundo de los datos de cada columna 

# Concatenar con tipo de via y nº para obtenter la dirección completa en solamente una columna
centros_medicos['NOMBRE-VIA'] 

# Concatenar con nombre de la vía via y nº para obtenter la dirección completa en solamente una columna
centros_medicos['CLASE-VIAL'] 

# Vamos a eliminar esta columna, información irrelevante para nuestro proyecto 
centros_medicos['TIPO-NUM'].unique()

# Concatenar con tipo de via y nombre de la vía para obtenter la dirección completa en solamente una columna
centros_medicos['NUM']

# Podemos eliminar esta columna ya que TODAS SON de madrid
centros_medicos['LOCALIDAD'].unique()

# Podemos eliminarla ya que TODAS SON de madrid
centros_medicos['PROVINCIA'].unique()

# QUITAR LOS CEROS
centros_medicos['CODIGO-POSTAL']

# Vamos a eliminar esta columna, nos vamos a quedr solmente con el distrito
centros_medicos['BARRIO']

centros_medicos['DISTRITO']

centros_medicos.TIPO[0]

#### Eliminación de las columnas con datos irrelevantes para el desarrollo de este proyecto

centros_medicos = centros_medicos.drop(['PK', 'TRANSPORTE', 'ACCESIBILIDAD', 'CONTENT-URL', 'TIPO-NUM', 'LOCALIDAD', 'PROVINCIA', 'BARRIO', 'TELEFONO'], axis=1)
centros_medicos.head()

centros_medicos.info()

#### Concatenamos columnas.
Pasos a seguir: 
    
    1. Primero cambiaremos los valores nulos por un cero
    2. Tranformaremos la columa de float a string
    3. Concatenaremos las columnas: Clase de via, Nombre de via y nº de via
    3. La del codigo postal aún no la junto porque tengo una idea (me gustaria que en la API con tan solo poner el codigo postas en vez de la latitud y longitud funcionara y te devolviera los tiempos, la distancia y la ruta)

centros_medicos.isnull().sum()

centros_medicos['NUMERO'] = centros_medicos['NUM'].fillna(0).astype(np.int64)
centros_medicos['NUMERO'] = centros_medicos['NUMERO'].astype(str)

centros_medicos['COD_POSTAL'] = centros_medicos['CODIGO-POSTAL'].fillna(0).astype(np.int64)
centros_medicos['COD_POSTAL'] = centros_medicos['COD_POSTAL'].astype(str)

centros_medicos = centros_medicos.drop(['NUM', 'CODIGO-POSTAL', 'Unnamed: 30'], axis=1)
centros_medicos.head()

centros_medicos.info()

centros_medicos['DIRECCION'] = centros_medicos['CLASE-VIAL'] + ' ' + centros_medicos['NOMBRE-VIA'] + ' ' + centros_medicos['NUMERO']

centros_medicos = centros_medicos.drop(['NOMBRE-VIA', 'CLASE-VIAL', 'NUMERO'], axis=1)
centros_medicos.head()

centros_medicos_clean = centros_medicos.reindex(columns=['NOMBRE','DIRECCION','DISTRITO','COD_POSTAL','COORDENADA-X', 'COORDENADA-Y', 'LATITUD', 'LONGITUD', 'TIPO'])
centros_medicos_clean.head()

#### LATITUD Y LONGITUD
Como podemos observar la latitud y longitud tiene una sintaxis NO válida para a posteriori plotear mapas y calcular distancias por lo que ahora necesitamos:

1. Eliminar todos los caratares(puntos) que haya entre medias de los numero que forman la longitud y latitud

2. Y luego, añadirle un punto entre la segunda posicion y tercera posicion de cada elemeno de la columna de Longitud y Latitud
  
   * ¡OJO! En la losngitud el simbolo menos (-) es necesario y cuenta como una posicion
   
   
3. Convertiremos estas dos columas a intengers para tener acabada nuestro análisis y limpieza de nuestro csv de centros medicos en madrid

A) LATITUD

centros_medicos_clean['LATITUD'] = centros_medicos_clean['LATITUD'].str.replace('.', '')

centros_medicos_clean['LATITUD'] = centros_medicos_clean['LATITUD'].apply(lambda x: x[:2] + '.' + x[2:])
centros_medicos_clean['LATITUD']

B) LONGITUD

centros_medicos_clean['LONGITUD'] = centros_medicos_clean['LONGITUD'].str.replace('.', '')

centros_medicos_clean['LONGITUD'] = centros_medicos_clean['LONGITUD'].apply(lambda x: x[:2] + '.' + x[2:])
centros_medicos_clean['LONGITUD']

#### TIPO
Los tipos de centros medicos si nos damos cuenta son nombres muy largos, por lo que vamos a inspeccionar un poco sobre ellos y ver si podemos acortarlos de alguna manera para poder diferenciarlos mejor.

centros_medicos_clean['TIPO']

centros_medicos_clean['TIPO'].unique()

* Voy a eliminar los primeros 54 caracteres porque en todos son los mismo y asi vamos a poder diferencias mucho mejor los tipos de centros medicos que hay en madrid.

centros_medicos_clean['TIPO'] = centros_medicos_clean['TIPO'].apply(lambda x: x[54:])
centros_medicos_clean['TIPO']

centros_medicos_clean['TIPO'].unique()

# Comprobamos si se han cambiado las longitudes y latitudes a como estabamos buscando
centros_medicos_clean.head()

# Guardamos centros_medicos_clean.csv en la carpeta de data
centros_medicos_clean.to_csv('../data/centros_medicos_clean.csv')

### 1.2 Farmacias en madrid

farmacias = pd.read_csv("../data/oficinas_farmacia.csv", encoding='latin-1', sep=';', error_bad_lines=False)
farmacias.head()

farmacias.shape

farmacias.keys()

farmacias.info()

farmacias.isnull().sum()

* En este caso nuestro dataframe no tiene apenas valores nulos en comparación con la cantidad de filas que tenemos. Por ello vamos a elimanar las columnas que tengas valores nulos.
* Lo que haremos en vez de eliminar columnas será concatenar las columnas de: 'direccion_vial_tipo', 'direccion_vial_nombre', 'direccion_vial_nro' porque si nos fijamos es el nombre de la calle partida y dividida en 3 columnas. 
* Al juntar estas tres columnas veremos si los valores NAN de las direcciones coinciden con los valores NAN de las columnas de las coordenadas (que son las que necesitamos)
    - Si estos valores NAN no coincidieraran eliminaremos solamente las filas que contengan valores nulos en las columnas de coordenadas X y coordenadas Y
    
    - Y, si si coinciden, lo que haremos será eliminar esas filas que al final y al cabo no va a afectar al desarrollo de este proyecto porque son una cantidad insignificante de datos de los cuales nos vamos a poder deshacer.

farmacias['direccion'] = farmacias['direccion_vial_tipo'] + ' ' + farmacias['direccion_vial_nombre'] + ' ' + farmacias['direccion_vial_nro']

farmacias = farmacias.drop(['direccion_vial_tipo', 'direccion_vial_nombre', 'direccion_vial_nro', 'farmacia_nro_soe'], axis=1)
farmacias.head()

farmacias = farmacias.reindex(columns=['titular_nombre','direccion','municipio_nombre','localizacion_coordenada_x','localizacion_coordenada_y'])
farmacias.head()

farmacias.isnull().sum()

# Investigamos en que filas tenemos nuestros valores nulos
nan_rows = farmacias[farmacias.isnull().any(1)]
nan_rows

* Las únicas filas que vamos a eliminar son aquellas que contengan los valores nulos en las coordenadas ya que es en lo que nos vamos a fijar aposteriori. Porque el no tener la dirección exacta de la las farmacias no nos perjudica al derrallo de este proyecto ya que poseemos sus coordenadas
* Esto lo haremos de la siguiente manera:

farmacias_clean = farmacias.dropna(subset=['localizacion_coordenada_x', 'localizacion_coordenada_y'])
farmacias_clean.head()

farmacias_clean.isnull().sum()

* Perfecto, ya tendriamos nuestro dataframe de farmcaias limpio y listo para trabjar con él.
* Lo guardaremos como un csv nuevo en nuestra carpeta de data para poder exportarlo en los siguientes notebooks y trabajar con él. de manera limpia y ordenada
* Lo guadaremos de la siguinete manera: 

# Guardamos farmacias_clean.csv en la carpeta de data
farmacias_clean.to_csv('../data/farmacias_clean.csv')
