==> notebooks/3.Calculo_de_distancias.ipynb <==
import pandas as pd
import math
import numpy as np

## Cálculo distacias: YourHome - YourHospital
### ¡ Vamos a ver cuánto de podrías ahorrar evitando ir a por una receta médica al Hospital!

* Calculamos la distacias entre el hostpiral y a tu casa

centros_medicos_clean = pd.read_csv("../data/centros_medicos_clean.csv")
centros_medicos_clean.head()

def haversine(coord1, coord2):
    R = 6372800  # Earth radius in meters
    lat1, lon1 = coord1
    lat2, lon2 = coord2
    
    phi1, phi2 = math.radians(lat1), math.radians(lat2) 
    dphi       = math.radians(lat2 - lat1)
    dlambda    = math.radians(lon2 - lon1)
    
    a = math.sin(dphi/2)**2 + \
        math.cos(phi1)*math.cos(phi2)*math.sin(dlambda/2)**2
    
    return 2*R*math.atan2(math.sqrt(a), math.sqrt(1 - a))

home = (40.4969434, -3.712521117)

for index, row in centros_medicos_clean.iterrows(): 
    row['LATITUD']
    row['LONGITUD']

centros_medicos_clean['DISTANCIA'] = 0 

centros_medicos_clean['DISTANCIA'] = centros_medicos_clean.apply(lambda row: haversine(home, (row['LATITUD'], row['LONGITUD'])), axis=1)

centros_medicos_clean.head()

centros_medicos_clean.loc[centros_medicos_clean.DISTANCIA.argmin(), "NOMBRE"]

centros_medicos_clean.loc[centros_medicos_clean.DISTANCIA.argmin(), "DIRECCION"]

# LA DISTANCIA ESTÁ CALCULADA EN MENTROS

centros_medicos_clean.loc[centros_medicos_clean.DISTANCIA.argmin(), "DISTANCIA"]

###  Funciones para aplicar en visual

def haversine(coord1, coord2):
    
    '''
    Funcion en donde al meterle dos parametros de coordenadas te va a devolver la distancia que hay entre ambos puntos
    '''
    
    R = 6372800  # Earth radius in meters
    lat1, lon1 = coord1
    lat2, lon2 = coord2
    
    phi1, phi2 = math.radians(lat1), math.radians(lat2) 
    dphi       = math.radians(lat2 - lat1)
    dlambda    = math.radians(lon2 - lon1)
    
    a = math.sin(dphi/2)**2 + \
        math.cos(phi1)*math.cos(phi2)*math.sin(dlambda/2)**2
    
    return 2*R*math.atan2(math.sqrt(a), math.sqrt(1 - a))

def distancia_entre_hospital_casa(lat, lon):
    '''
    Esta funcion lo que coge es la latitud y longitud de la ubicacion de tu casa y te evuelve el nombre del centro medico 
    mas cercano a tu casa
    '''
    
    for index, row in centros_medicos_clean.iterrows(): 
        row['LATITUD']
        row['LONGITUD']
    
    centros_medicos_clean['DISTANCIA'] = 0 
    
    centros_medicos_clean['DISTANCIA'] = centros_medicos_clean.apply(lambda row: haversine((lat, lon), (row['LATITUD'], row['LONGITUD'])), axis=1)
    
    return (centros_medicos_clean.loc[centros_medicos_clean.DISTANCIA.argmin(), "NOMBRE"], math.trunc(centros_medicos_clean.loc[centros_medicos_clean.DISTANCIA.argmin(), "DISTANCIA"]))
    

distancia_entre_hospital_casa(40.4969434, -3.712521117)

# PRUEBA CON LA DISTANCIAS QUE ME HA DADO GEOCODE (no es tan excata porque las coordenadas tienen menos decimales, salen redondeadas por defecto)
# Pero acierta!

distancia_entre_hospital_casa(40.4969, -3.71037)

# SEGUNA PRUEBA CON LAS COORDENADAS QUE ME DEVUELVE LA API
distancia_entre_hospital_casa(40.4687, -3.68553)

## Cálculo distacias: Home-Pharmacies
### ¡Vamos a encontrar las farmacias más cercana a tu casa!

def haversine(coord1, coord2):
    R = 6372800  # Earth radius in meters
    lat1, lon1 = coord1
    lat2, lon2 = coord2
    
    phi1, phi2 = math.radians(lat1), math.radians(lat2) 
    dphi       = math.radians(lat2 - lat1)
    dlambda    = math.radians(lon2 - lon1)
    
    a = math.sin(dphi/2)**2 + \
        math.cos(phi1)*math.cos(phi2)*math.sin(dlambda/2)**2
    
    return 2*R*math.atan2(math.sqrt(a), math.sqrt(1 - a))

farmacias_clean = pd.read_csv("../data/farmacias_clean.csv")
farmacias_clean.head()

home = (40.4969434, -3.712521117)

* La columna de DISTACIA será la distancia que hay entre tus casa y todas las farmacias de Madrid
* Luego encontraremos la farmacia más cercana a tu casa

farmacias_clean['DISTANCIA'] = 0 

farmacias_clean.head()

for index, row in farmacias_clean.iterrows(): 
    row['latitud']
    row['longitud']

farmacias_clean['DISTANCIA'] = farmacias_clean.apply(lambda row: haversine(home, (row['latitud'], row['longitud'])), axis=1)

farmacias_clean.head()

farmacias_clean.loc[farmacias_clean.DISTANCIA.argmin(), "DISTANCIA"]

farmacias_clean.loc[farmacias_clean.DISTANCIA.argmin(), "direccion"]

farmacias_clean.loc[farmacias_clean.DISTANCIA.argmin(), "titular_nombre"]

### Funciones para calcular la distacias de tu casa a la farmcaia más cercana (el recorrido que va a hacer el repartidor por ti)


def distancia_entre_farmacia_casa(lat, lon):
    for index, row in farmacias_clean.iterrows(): 
        row['latitud']
        row['longitud']
    
    farmacias_clean['DISTANCIA'] = 0 
    
    farmacias_clean['DISTANCIA'] = farmacias_clean.apply(lambda row: haversine(home, (row['latitud'], row['longitud'])), axis=1)
    
    return (farmacias_clean.loc[farmacias_clean.DISTANCIA.argmin(), "titular_nombre"], farmacias_clean.loc[farmacias_clean.DISTANCIA.argmin(), "DISTANCIA"])

distancia_entre_farmacia_casa(40.4969434, -3.712521117)

# ACIERTAAAA!! SON LAS COORDENADAS QUE ME DA MI API AL METER YO MI CALLE
distancia_entre_farmacia_casa(40.4969, -3.71037)

distancia_entre_farmacia_casa(40.4687, -3.68553)
