==> notebooks/4.ApiGeocode.ipynb <==
import requests
import json
from dotenv import load_dotenv
import os
import pandas as pd


import pandas as pd
import math
import numpy as np


import folium
from folium import Choropleth, Circle, Marker, Icon, Map
from folium import plugins
from folium.plugins import HeatMap, MarkerCluster
from folium import IFrame
import ipywidgets
from folium.features import CustomIcon

## API GEOCODE
### Añadiendo el key ID para poder utilizar la API

load_dotenv()

key = os.getenv("key")

### Función que va a convertir tu direccion en latitud y longitud

def geocode(calle):
    """
    Saca las coordenadas de una dirección que le des.
    """
    data = requests.get(f"https://geocode.xyz/{calle}?json=1").json()
    try:
        casa = [float(data["latt"]),float(data["longt"])]
        return casa
    except:
        return data

calle = "Calle la Maso 1 28034 Madrid"

geocode(calle)

#### Primer ejemplo

calle = "Calle Marbella 83 28034, Madrid"

data = requests.get(f"https://geocode.xyz/{calle}?json=1").json()
data

home = geocode(calle)
home

home.keys()

#### Segundo ejemplo

calle = "Calle Nuñez Morgado 9, Madrid"

geocode(calle)

####  Tercer ejemplo

calle = "Calle de Villanueva, 11 29001 Madrid"

coordenadas = geocode(calle)
coordenadas

## JUGANDO CON FUNCIONES

### Calculamos con la calle dada por el paciente el centro medico más cercano

- La calle dada es en froma de string. Este string recorreria nuestra funcion formada por la API de Geocode.
- Posteriormente las coordenadas sacadas se introcuden en la funcion de calculo con todos los centros medicos y te devuelve el centro más cercano a tu casa

centros_medicos_clean = pd.read_csv("../data/centros_medicos_clean.csv")
centros_medicos_clean.head()

calle = "Calle Tembleque, 52 28024 Madrid"

def geocode(calle):
    """
    Saca las coordenadas de una dirección que le des.
    """
    data = requests.get(f"https://geocode.xyz/{calle}?json=1").json()
    try:
        casa = [float(data["latt"]),float(data["longt"])]
        return casa
    except:
        return data

casa = geocode(calle)
casa

def haversine(coord1, coord2):
    
    '''
    Funcion en donde al meterle dos parametros de coordenadas te va a devolver la distancia que hay entre ambos puntos
    '''
    
    R = 6372800  # Earth radius in meters
    lat1, lon1 = coord1
    lat2, lon2 = coord2
    
    phi1, phi2 = math.radians(lat1), math.radians(lat2) 
    dphi       = math.radians(lat2 - lat1)
    dlambda    = math.radians(lon2 - lon1)
    
    a = math.sin(dphi/2)**2 + \
        math.cos(phi1)*math.cos(phi2)*math.sin(dlambda/2)**2
    
    return 2*R*math.atan2(math.sqrt(a), math.sqrt(1 - a))

def distancia_entre_hospital_casa(casa):

    '''
    Esta funcion lo que coge es la latitud y longitud de la ubicacion de tu casa y te evuelve el nombre del centro medico 
    mas cercano a tu casa
    '''
    
    for index, row in centros_medicos_clean.iterrows(): 
        row['LATITUD']
        row['LONGITUD']
    
    centros_medicos_clean['DISTANCIA'] = 0 
    
    centros_medicos_clean['DISTANCIA'] = centros_medicos_clean.apply(lambda row: haversine((casa), (row['LATITUD'], row['LONGITUD'])), axis=1)
    
    return (centros_medicos_clean.loc[centros_medicos_clean.DISTANCIA.argmin(), "NOMBRE"], math.trunc(centros_medicos_clean.loc[centros_medicos_clean.DISTANCIA.argmin(), "DISTANCIA"]))
    

distancia_entre_hospital_casa(casa)

###  Cálculo de la farmacia más cercana a tu casa

- Introducioendo el strind de la dirreccion de tu casa en la funcion creada por la API geocode sacaremos las coordenadas de tu casa.
- Posterirmente estas coordenadas las meteremos en otra función calculando asi la distancia con todas las farmacias de Madrid y devolviendo la farmacia más cercana a tu casa

farmacias_clean = pd.read_csv("../data/farmacias_clean.csv")
farmacias_clean.head()

casa = geocode(calle)
casa

def distancia_entre_farmacia_casa(casa):
    for index, row in farmacias_clean.iterrows(): 
        row['latitud']
        row['longitud']
    
    farmacias_clean['DISTANCIA'] = 0 
    
    farmacias_clean['DISTANCIA'] = farmacias_clean.apply(lambda row: haversine(casa, (row['latitud'], row['longitud'])), axis=1)
    
    return (farmacias_clean.loc[farmacias_clean.DISTANCIA.argmin(), "titular_nombre"], farmacias_clean.loc[farmacias_clean.DISTANCIA.argmin(), "DISTANCIA"])

distancia_entre_farmacia_casa(casa)

### Mapa para visualizar las 3 localizaciones recibidas:
    - Direccioón de tu casa
    - Dirección de Hospital más cercano
    - Dirección de la farmacia más cercana

#### Función que saca las coordenadas de la farmacia más cercana a la dirreción que se haya introducido

def coordenadas_farmacia_mas_cercana(casa):
    for index, row in farmacias_clean.iterrows(): 
        row['latitud']
        row['longitud']
     
    farmacias_clean['DISTANCIA'] = 0
    farmacias_clean['DISTANCIA'] = farmacias_clean.apply(lambda row: haversine(casa, (row['latitud'], row['longitud'])), axis=1)
    
    return (farmacias_clean.loc[farmacias_clean.DISTANCIA.argmin(), "latitud"], farmacias_clean.loc[farmacias_clean.DISTANCIA.argmin(), "longitud"])

coord_farm_cercana = coordenadas_farmacia_mas_cercana(casa)
coord_farm_cercana

#### Función que saca las coordenadas del hospital más cercano a la dirección que se haya introducido

def coordenadas_hospital_mas_cercano(casa):

    '''
    Esta funcion lo que coge es la latitud y longitud de la ubicacion de tu casa y te evuelve el nombre del centro medico 
    mas cercano a tu casa
    '''
    
    for index, row in centros_medicos_clean.iterrows(): 
        row['LATITUD']
        row['LONGITUD']
    
    centros_medicos_clean['DISTANCIA'] = 0 
    
    centros_medicos_clean['DISTANCIA'] = centros_medicos_clean.apply(lambda row: haversine((casa), (row['LATITUD'], row['LONGITUD'])), axis=1)
    
    return (centros_medicos_clean.loc[centros_medicos_clean.DISTANCIA.argmin(), "LATITUD"], (centros_medicos_clean.loc[centros_medicos_clean.DISTANCIA.argmin(), "LONGITUD"]))
    

coord_hospital_cercano = coordenadas_hospital_mas_cercano(casa)
coord_hospital_cercano

#### Mapa con las tres variables de coordenadas sacadas

# VARIABLES DE LOS ICONOS (en visual no tienes que ponerlos porque ya estan definidos)
path = "../icons/{}".format

icon_hosp = path("Hospital-clinica-sanatorio-icon.png")
icon_farm = path("cruz-icon.png")

# Las printeo solamente para ver que no hay ningun error
print(casa)
print(coord_farm_cercana)
print(coord_hospital_cercano)

#### PRUEBAS VISUALES

##### Con tres variables (tu casa, la famarcias más cercana y el centro medico mas cercano)

map_final = Map(location = casa, zoom_start=15, control_scale = True)

def distancia_visual(map_final):
    map_final = Map(location = casa, zoom_start=15, control_scale = True)
    #HOSPiTAL
    icon = CustomIcon( 
                icon_hosp, 
                icon_size=(40, 50))
    loc = coord_hospital_cercano

    hospital_cercano = Marker(location = loc, icon = icon, tooltip = "El hospital mas cercano")
    hospital_cercano.add_to(map_final)


    #FARMACIA
    icon = CustomIcon(
                icon_farm,
                icon_size=(40, 50))

    loc = coord_farm_cercana
    farmacia_cercana = Marker(location = loc,icon = icon, tooltip = "La farmacia mas cercana")
    farmacia_cercana.add_to(map_final)


    #HOME
    icon = Icon(color = "red",
                        prefix = "fa",
                        icon = "home",
                        icon_color = "white",
                        icon_size = (40, 50))
    loc = casa
    mi_casa = Marker(location = loc, icon = icon, tooltip = "Tu casa")
    mi_casa.add_to(map_final)


    return map_final

distancia_visual(map_final)

##### Con dos variables (tu casa y la farmcacia más cercana)

map_final_2 = Map(location = casa, zoom_start=15, control_scale = True)

def distancia_visual(map_final_2):
    
    #FARMACIA
    icon = CustomIcon(
                icon_farm,
                icon_size=(40, 50))

    loc = coord_farm_cercana
    farmacia_cercana = Marker(location = loc,icon = icon, tooltip = "La farmacia mas cercana")
    farmacia_cercana.add_to(map_final_2)


    #HOME
    icon = Icon(color = "red",
                        prefix = "fa",
                        icon = "home",
                        icon_color = "white",
                        icon_size = (40, 50))
    loc = casa
    mi_casa = Marker(location = loc, icon = icon, tooltip = "Tu casa")
    mi_casa.add_to(map_final_2)


    return map_final_2

distancia_visual(map_final_2)
