==> notebooks/5.Documentacion.ipynb <==
#  Documentacion 

## Dataset's resources

* BORRAR --> dataset de hospitales de españa : https://opendata.esri.es/datasets/68745a7fb7a348b6b0d722c8517790af_0/data


* dataset de centros medicos en madrid: https://datos.madrid.es/portal/site/egob/menuitem.c05c1f754a33a9fbe4b2e4b284f1a5a0/?vgnextoid=da7437ac37efb410VgnVCM2000000c205a0aRCRD&vgnextchannel=374512b9ace9f310VgnVCM100000171f5a0aRCRD&vgnextfmt=default


* pdf de la documentacion del dataset de los centros medicos en madrid: https://datos.madrid.es/FWProjects/egob/Catalogo/XComun/Ficheros/Estructura_DS_ConjuntoDatos.pdf


* dataset de farmacias en madrid : https://datos.comunidad.madrid/catalogo/dataset/oficinas_farmacia

## Api Geocode

Documentacion de Geocode: https://geocode.xyz/

## Libreraries resources

Documentacion libreria pandas: 
https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.Series.unique.html

Documentacion libreria de folium : 

Documentacion libreria streamlit: https://docs.streamlit.io/en/stable/getting_started.html
